+++ internal/controller/grafanaorganization_controller.go
@@ -23,16 +23,24 @@ import (
 	grafanaAPI "github.com/grafana/grafana-openapi-client-go/client"
 	grafanaAPIModels "github.com/grafana/grafana-openapi-client-go/models"
 	"github.com/pkg/errors"
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/cluster-api/util/patch"
 	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/giantswarm/observability-operator/api/v1alpha1"
+	"github.com/giantswarm/observability-operator/internal/controller/predicates"
+	"github.com/giantswarm/observability-operator/pkg/grafana"
+	"github.com/giantswarm/observability-operator/pkg/grafana/templating"
 )
-
-const sharedOrgName = "Shared Org."
 
 // GrafanaOrganizationReconciler reconciles a GrafanaOrganization object
 type GrafanaOrganizationReconciler struct {
@@ -62,16 +70,6 @@ func (r *GrafanaOrganizationReconciler) Reconcile(ctx context.Context, req ctrl.
 		return ctrl.Result{}, errors.WithStack(client.IgnoreNotFound(err))
 	}
 
-	// Test connection to Grafana
-	// TODO(zirko) Remove in the next iteration
-	_, err = r.GrafanaAPI.Health.GetHealth()
-	if err != nil {
-		logger.Error(err, "Failed to connect to Grafana")
-		return ctrl.Result{}, errors.WithStack(err)
-	}
-
-	logger.Info("Successfully connected to Grafana, lets start hacking...")
-
 	// Handle deleted grafana organizations
 	if !grafanaOrganization.DeletionTimestamp.IsZero() {
 		return ctrl.Result{}, r.reconcileDelete(ctx, grafanaOrganization)
@@ -82,30 +80,76 @@ func (r *GrafanaOrganizationReconciler) Reconcile(ctx context.Context, req ctrl.
 }
 
 // reconcileCreate creates the grafanaOrganization.
+// reconcileCreate ensures the Grafana organization described in grafanaOrganization CR is created in Grafana.
+// This function is also responsible for:
+// - Adding the finalizer to the CR
+// - Updating the CR status field
+// - Renaming the Grafana Main Org.
 func (r GrafanaOrganizationReconciler) reconcileCreate(ctx context.Context, grafanaOrganization *v1alpha1.GrafanaOrganization) (ctrl.Result, error) { // nolint:unparam
 	logger := log.FromContext(ctx)
 
-	originalGrafanaOrganization := grafanaOrganization.DeepCopy()
-	// If the grafanaOrganization doesn't have our finalizer, add it.
-	if controllerutil.AddFinalizer(grafanaOrganization, v1alpha1.GrafanaOrganizationFinalizer) {
-		logger.Info("Add finalizer to Grafana Organization")
-		// Register the finalizer immediately to avoid orphaning resources on delete
-		if err := r.Client.Patch(ctx, grafanaOrganization, client.MergeFrom(originalGrafanaOrganization)); err != nil {
+	// Add finalizer first if not set to avoid the race condition between init and delete.
+	if !controllerutil.ContainsFinalizer(grafanaOrganization, v1alpha1.GrafanaOrganizationFinalizer) {
+		// We use a patch rather than an update to avoid conflicts when multiple controllers are adding their finalizer to the grafana organization
+		// We use the patch from sigs.k8s.io/cluster-api/util/patch to handle the patching without conflicts
+		logger.Info("adding finalizer", "finalizer", v1alpha1.GrafanaOrganizationFinalizer)
+		patchHelper, err := patch.NewHelper(grafanaOrganization, r.Client)
+		if err != nil {
 			return ctrl.Result{}, errors.WithStack(err)
 		}
+		controllerutil.AddFinalizer(grafanaOrganization, v1alpha1.GrafanaOrganizationFinalizer)
+		if err := patchHelper.Patch(ctx, grafanaOrganization); err != nil {
+			logger.Error(err, "failed to add finalizer", "finalizer", v1alpha1.GrafanaOrganizationFinalizer)
+			return ctrl.Result{}, errors.WithStack(err)
+		}
+		logger.Info("added finalizer", "finalizer", v1alpha1.GrafanaOrganizationFinalizer)
+		return ctrl.Result{}, nil
 	}
 
 	// Ensure the first organization is renamed.
 	_, err := r.GrafanaAPI.Orgs.UpdateOrg(1, &grafanaAPIModels.UpdateOrgForm{
-		Name: sharedOrgName,
+		Name: grafana.SharedOrgName,
 	})
 	if err != nil {
-		logger.Error(err, fmt.Sprintf("Could not rename Main Org. to %s", sharedOrgName))
+		logger.Error(err, fmt.Sprintf("failed to rename Main Org. to %s", grafana.SharedOrgName))
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
-	//TODO Implement the logic to create the Grafana organization
+	// TODO add datasources for shared org.
 
+	// Create or update organization in Grafana
+	var organization = grafana.Organization{
+		ID:   grafanaOrganization.Status.OrgID,
+		Name: grafanaOrganization.Spec.DisplayName,
+	}
+
+	if organization.ID == 0 {
+		// if the CR doesn't have an orgID, create the organization in Grafana
+		organization, err = grafana.CreateOrganization(ctx, r.GrafanaAPI, organization)
+	} else {
+		organization, err = grafana.UpdateOrganization(ctx, r.GrafanaAPI, organization)
+	}
+
+	if err != nil {
+		return ctrl.Result{}, err
+	}
+
+	// Update CR status if anything was changed
+	if organization.ID != grafanaOrganization.Status.OrgID {
+		grafanaOrganization.Status.OrgID = organization.ID
+
+		if err = r.Status().Update(ctx, grafanaOrganization); err != nil {
+			logger.Error(err, "failed to update grafanaOrganization status")
+			return ctrl.Result{}, errors.WithStack(err)
+		}
+	}
+
+	// TODO add datasources for the organization.
+
+	err = r.configureGrafana(ctx)
+	if err != nil {
+		return ctrl.Result{}, errors.WithStack(err)
+	}
 	return ctrl.Result{}, nil
 }
 
@@ -113,17 +157,44 @@ func (r GrafanaOrganizationReconciler) reconcileCreate(ctx context.Context, graf
 func (r GrafanaOrganizationReconciler) reconcileDelete(ctx context.Context, grafanaOrganization *v1alpha1.GrafanaOrganization) error {
 	logger := log.FromContext(ctx)
 
-	//TODO Implement the logic to delete the organization from Grafana.
+	// We do not need to delete anything if there is no finalizer on the grafana organization
+	if !controllerutil.ContainsFinalizer(grafanaOrganization, v1alpha1.GrafanaOrganizationFinalizer) {
+		return nil
+	}
 
-	logger.Info("Remove finalizer from grafana organization")
-	// Remove the finalizer.
-	originalGrafanaOrganization := grafanaOrganization.DeepCopy()
-	if controllerutil.RemoveFinalizer(grafanaOrganization, v1alpha1.GrafanaOrganizationFinalizer) {
-		err := r.Client.Patch(ctx, grafanaOrganization, client.MergeFrom(originalGrafanaOrganization))
+	// Delete organization in Grafana if it exists
+	if grafanaOrganization.Status.OrgID > 0 {
+		err := grafana.DeleteByID(ctx, r.GrafanaAPI, grafanaOrganization.Status.OrgID)
 		if err != nil {
-			return err
+			return errors.WithStack(err)
+		}
+
+		grafanaOrganization.Status.OrgID = 0
+		if err = r.Status().Update(ctx, grafanaOrganization); err != nil {
+			logger.Error(err, "failed to update grafanaOrganization status")
+			return errors.WithStack(err)
 		}
 	}
+
+	err := r.configureGrafana(ctx)
+	if err != nil {
+		return errors.WithStack(err)
+	}
+
+	// Finalizer handling needs to come last.
+	// We use the patch from sigs.k8s.io/cluster-api/util/patch to handle the patching without conflicts
+	logger.Info("removing finalizer", "finalizer", v1alpha1.GrafanaOrganizationFinalizer)
+	patchHelper, err := patch.NewHelper(grafanaOrganization, r.Client)
+	if err != nil {
+		return errors.WithStack(err)
+	}
+
+	controllerutil.RemoveFinalizer(grafanaOrganization, v1alpha1.GrafanaOrganizationFinalizer)
+	if err := patchHelper.Patch(ctx, grafanaOrganization); err != nil {
+		logger.Error(err, "failed to remove finalizer, requeuing", "finalizer", v1alpha1.GrafanaOrganizationFinalizer)
+		return errors.WithStack(err)
+	}
+	logger.Info("removed finalizer", "finalizer", v1alpha1.GrafanaOrganizationFinalizer)
 
 	return nil
 }
@@ -132,5 +203,80 @@ func (r GrafanaOrganizationReconciler) reconcileDelete(ctx context.Context, graf
 func (r *GrafanaOrganizationReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&v1alpha1.GrafanaOrganization{}).
+		// Watch for grafana pod's status changes
+		Watches(
+			&v1.Pod{},
+			handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, obj client.Object) []reconcile.Request {
+				k8sClient := mgr.GetClient()
+				var organizations v1alpha1.GrafanaOrganizationList
+
+				err := k8sClient.List(ctx, &organizations)
+				if err != nil {
+					log.FromContext(ctx).Error(err, "failed to list grafana organization CRs")
+					return []reconcile.Request{}
+				}
+
+				// Reconcile all grafana organizations when the grafana pod is recreated
+				requests := make([]reconcile.Request, 0, len(organizations.Items))
+				for _, organization := range organizations.Items {
+					requests = append(requests, reconcile.Request{
+						NamespacedName: types.NamespacedName{
+							Name: organization.Name,
+						},
+					})
+				}
+				return requests
+			}),
+			builder.WithPredicates(predicates.GrafanaPodRecreatedPredicate{}),
+		).
 		Complete(r)
+}
+
+// configureGrafana ensures the RBAC configuration is set in Grafana.
+func (r *GrafanaOrganizationReconciler) configureGrafana(ctx context.Context) error {
+	logger := log.FromContext(ctx)
+
+	organizations := v1alpha1.GrafanaOrganizationList{}
+	err := r.Client.List(ctx, &organizations)
+	if err != nil {
+		logger.Error(err, "failed to list grafana organizations.")
+		return errors.WithStack(err)
+	}
+
+	grafanaConfig := v1.ConfigMap{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "grafana-user-values",
+			Namespace: "giantswarm",
+		},
+	}
+
+	_, err = controllerutil.CreateOrPatch(ctx, r.Client, &grafanaConfig, func() error {
+		config, err := templating.GenerateGrafanaConfiguration(organizations.Items)
+		if err != nil {
+			logger.Error(err, "failed to generate grafana user configmap values.")
+			return errors.WithStack(err)
+		}
+
+		for _, organization := range organizations.Items {
+			// Set owner reference to the config map to be able to clean it up when all organizations are deleted
+			err = controllerutil.SetOwnerReference(&organization, &grafanaConfig, r.Scheme)
+			if err != nil {
+				return errors.WithStack(err)
+			}
+		}
+
+		logger.Info("updating grafana-user-values", "config", config)
+
+		grafanaConfig.Data = make(map[string]string)
+		grafanaConfig.Data["values"] = config
+
+		return nil
+	})
+
+	if err != nil {
+		logger.Error(err, "failed to configure grafana.")
+		return errors.WithStack(err)
+	}
+
+	return nil
 }